+++ examples/cluster2.lua
@@ -4,6 +4,5 @@ local cluster = require "cluster"
 skynet.start(function()
 	local proxy = cluster.proxy("db", ".simpledb")
 	print(skynet.call(proxy, "lua", "GET", "a"))
-	print(cluster.ncall("db.simpledb","GET", "a"))
 	print(cluster.call("db", ".simpledb", "GET", "a"))
 end)