+++ HISTORY.md
@@ -8,7 +8,7 @@ Dev version
 * snax service add change SERVICE_PATH and add it to package.path
 * skynet.redirect support string address
 * add skynet.harbor.queryname to query globalname
-* add cluster.proxy and cluster.ncall
+* add cluster.proxy 
 
 v0.5.2 (2014-8-11)
 -----------